+++ process.c
@@ -6,7 +6,7 @@
 proc os_ptable[NPROCS];
 unsigned os_nprocs = 0;
 
-unsigned long long score(proc p){return p->n_ticks/p->n_calls;}
+unsigned long score(proc p){return p->n_ticks/p->n_calls;}
 
 /* create_proc -- allocate and initialise process descriptor */
 proc create_proc(char *name, unsigned stksize) {
